Refactor MediaType enum to separate class
