+++ android/src/main/java/com/thebluealliance/androidclient/Constants.java
@@ -1,6 +1,6 @@
 package com.thebluealliance.androidclient;
 
-import com.thebluealliance.androidclient.models.Media;
+import com.thebluealliance.androidclient.helpers.MediaType;
 
 import java.util.HashMap;
 
@@ -19,7 +19,7 @@ public class Constants {
     public static final String LAST_YEAR_KEY = "last_season";
 
     public static final HashMap<String, String> MATCH_LEVELS;
-    public static final HashMap<Media.TYPE, String> MEDIA_IMG_URL_PATTERN,
+    public static final HashMap<MediaType, String> MEDIA_IMG_URL_PATTERN,
             MEDIA_LINK_URL_PATTERN;
 
     static {
@@ -39,10 +39,10 @@ public class Constants {
 
         MEDIA_IMG_URL_PATTERN = new HashMap<>();
         MEDIA_LINK_URL_PATTERN = new HashMap<>();
-        MEDIA_IMG_URL_PATTERN.put(Media.TYPE.CD_PHOTO_THREAD, "http://www.chiefdelphi.com/media/img/%s");
-        MEDIA_LINK_URL_PATTERN.put(Media.TYPE.CD_PHOTO_THREAD, "http://www.chiefdelphi.com/media/photos/%s");
-        MEDIA_IMG_URL_PATTERN.put(Media.TYPE.YOUTUBE, "http://img.youtube.com/vi/%s/hqdefault.jpg");
-        MEDIA_LINK_URL_PATTERN.put(Media.TYPE.YOUTUBE, "https://www.youtube.com/watch?v=%s");
+        MEDIA_IMG_URL_PATTERN.put(MediaType.CD_PHOTO_THREAD, "http://www.chiefdelphi.com/media/img/%s");
+        MEDIA_LINK_URL_PATTERN.put(MediaType.CD_PHOTO_THREAD, "http://www.chiefdelphi.com/media/photos/%s");
+        MEDIA_IMG_URL_PATTERN.put(MediaType.YOUTUBE, "http://img.youtube.com/vi/%s/hqdefault.jpg");
+        MEDIA_LINK_URL_PATTERN.put(MediaType.YOUTUBE, "https://www.youtube.com/watch?v=%s");
     }
 
     public static String getApiHeader() {
