+++ android/src/main/java/com/thebluealliance/androidclient/models/Media.java
@@ -8,40 +8,12 @@ import com.thebluealliance.androidclient.Constants;
 import com.thebluealliance.androidclient.database.Database;
 import com.thebluealliance.androidclient.database.tables.MediasTable;
 import com.thebluealliance.androidclient.helpers.JSONHelper;
+import com.thebluealliance.androidclient.helpers.MediaType;
 import com.thebluealliance.androidclient.listitems.ImageListElement;
 import com.thebluealliance.androidclient.listitems.ListElement;
 
 
 public class Media extends BasicModel<Media> {
-
-    public enum TYPE {
-        NONE,
-        YOUTUBE,
-        CD_PHOTO_THREAD;
-
-        public static TYPE fromString(String string) {
-            switch (string) {
-                case "cdphotothread":
-                    return CD_PHOTO_THREAD;
-                case "youtube":
-                    return YOUTUBE;
-                default:
-                    return NONE;
-            }
-        }
-
-        public String toString() {
-            switch (this) {
-                case NONE:
-                    return "";
-                case YOUTUBE:
-                    return "youtube";
-                case CD_PHOTO_THREAD:
-                    return "cdphotothread";
-            }
-            return "";
-        }
-    }
 
     private JsonObject details;
 
@@ -59,9 +31,9 @@ public class Media extends BasicModel<Media> {
         }
     }
 
-    public Media.TYPE getMediaType() throws FieldNotDefinedException {
+    public MediaType getMediaType() throws FieldNotDefinedException {
         if (fields.containsKey(MediasTable.TYPE) && fields.get(MediasTable.TYPE) instanceof String) {
-            return TYPE.fromString((String) fields.get(MediasTable.TYPE));
+            return MediaType.fromString((String) fields.get(MediasTable.TYPE));
         }
         throw new FieldNotDefinedException("Field Database.Medias.TYPE is not defined");
     }
@@ -70,7 +42,7 @@ public class Media extends BasicModel<Media> {
         fields.put(MediasTable.TYPE, typeString);
     }
 
-    public void setMediaType(Media.TYPE mediaType) {
+    public void setMediaType(MediaType mediaType) {
         fields.put(MediasTable.TYPE, mediaType.toString());
     }
 
@@ -131,17 +103,17 @@ public class Media extends BasicModel<Media> {
     public ListElement render() {
         String imageUrl;
         try {
-            TYPE mediaType = getMediaType();
+            MediaType mediaType = getMediaType();
             String foreignKey = getForeignKey();
-            if (mediaType == TYPE.CD_PHOTO_THREAD) {
+            if (mediaType == MediaType.CD_PHOTO_THREAD) {
                 JsonObject details = getDetails();
                 imageUrl = String.format(Constants.MEDIA_IMG_URL_PATTERN.get(mediaType), details.get("image_partial").getAsString().replace("_l.jpg", "_m.jpg"));
-            } else if (mediaType == TYPE.YOUTUBE) {
+            } else if (mediaType == MediaType.YOUTUBE) {
                 imageUrl = String.format(Constants.MEDIA_IMG_URL_PATTERN.get(mediaType), foreignKey);
             } else {
                 imageUrl = "";
             }
-            Boolean isVideo = mediaType == TYPE.YOUTUBE;
+            Boolean isVideo = mediaType == MediaType.YOUTUBE;
             return new ImageListElement(imageUrl,
                     String.format(Constants.MEDIA_LINK_URL_PATTERN.get(mediaType), foreignKey), isVideo);
         } catch (FieldNotDefinedException e) {
